chore(core): rename memory to mem
