+++ crates/core/src/memory/mem.rs
